==> code/04. Extract technical indicator.ipynb <==
from talib import *
import pandas as pd
import numpy as np

NASDAQ = pd.read_csv("../data/Stock_info/NASDAQ.csv")
NASDAQ.sort_values(by=["Market Cap"], ascending=False, inplace=True)
NASDAQ_symbols = NASDAQ[(NASDAQ["IPO Year"]<2009).values & (NASDAQ["IPO Year"].notna()).values & (NASDAQ["Name"].str.contains("Common Stock").values)][0:100]["Symbol"].values

NASDAQ_stock = pd.read_csv("../data/Stock_data/pre_NASDAQ.csv", encoding="cp949")

NASDAQ_stock.replace(0, np.NaN, inplace = True)

NASDAQ_stock.fillna(method="pad", inplace = True)

def TI_ex(symbol, OHLCV_data):
    Date = OHLCV_data["Date"]
    Open = OHLCV_data[symbol+"_Open"]
    High = OHLCV_data[symbol+"_High"]
    Low = OHLCV_data[symbol+"_Low"]
    Close = OHLCV_data[symbol+"_Close"]
    Volume = OHLCV_data[symbol+"_Volume"]
    
    #Overlap Studies Functions
    upperband, middleband, lowerband = BBANDS(Close, timeperiod=5, nbdevup=2, nbdevdn=2, matype=0)
    dema = DEMA(Close, timeperiod=30)
    ema = EMA(Close, timeperiod=30)
    ht_trendline = HT_TRENDLINE(Close)
    kama = KAMA(Close, timeperiod=30)
    ma = MA(Close, timeperiod=30, matype=0)
    midpoint = MIDPOINT(Close, timeperiod=14)
    midprice = MIDPRICE(High, Low, timeperiod=14)
    sma = SMA(Close, timeperiod=30)
    t3 = T3(Close, timeperiod=5, vfactor=0)
    tema = TEMA(Close, timeperiod=30)
    trima = TRIMA(Close, timeperiod=30)
    wma = WMA(Close, timeperiod=30)
    
    #Momentum Indicator Functions
    adx = ADX(High, Low, Close, timeperiod=14)
    apo = APO(Close, fastperiod=12, slowperiod=26, matype=0)
    aroondown, aroonup = AROON(High, Low, timeperiod=14)
    bop = BOP(Open, High, Low, Close)
    cci = CCI(High, Low, Close, timeperiod=14)
    cmo = CMO(Close, timeperiod=14)
    dx = DX(High, Low, Close, timeperiod=14)
    mfi = MFI(High, Low, Close, Volume, timeperiod=14)
    minus_di = MINUS_DI(High, Low, Close, timeperiod=14)
    mom = MOM(Close, timeperiod=10)
    plus_di = PLUS_DI(High, Low, Close, timeperiod=14)
    ppo = PPO(Close, fastperiod=12, slowperiod=26, matype=0)
    roc = ROC(Close, timeperiod=10)
    rsi = RSI(Close, timeperiod=14)
    slowk, slowd = STOCH(High, Low, Close, fastk_period=5, slowk_period=3, slowk_matype=0, slowd_period=3, slowd_matype=0)
    fastk, fastd = STOCHF(High, Low, Close, fastk_period=5, fastd_period=3, fastd_matype=0)
    trix = TRIX(Close, timeperiod=30)
    ultosc = ULTOSC(High, Low, Close, timeperiod1=7, timeperiod2=14, timeperiod3=28)
    willr = WILLR(High, Low, Close, timeperiod=14)
    
    #Volume Indicator Functions
    ad = AD(High, Low, Close, Volume)
    obv = OBV(Close, Volume)
    
    #Volatility Indicator Functions
    atr = ATR(High, Low, Close, timeperiod=14)
    trange = TRANGE(High, Low, Close)
    
    #Statistic Functions
    beta = BETA(High, Low, timeperiod=5)
    correl = CORREL(High, Low, timeperiod=30)
    linearreg = LINEARREG(Close, timeperiod=14)
    stddev = STDDEV(Close, timeperiod=5, nbdev=1)
    tsf = TSF(Close, timeperiod=14)
    var = VAR(Close, timeperiod=5, nbdev=1)
    
    
    TI_DF = pd.DataFrame({
        "Date":Date,
        "Open":Open,
        "Close":Close,
        "upperband":upperband,
        "middleband":middleband,
        "lowerband":lowerband,
        "dema":dema,
        "ema":ema,
        "ht_trendline":ht_trendline,
        "kama":kama,
        "ma":ma,
        "midpoint":midpoint,
        "midprice":midprice,
        "sma":sma,
        "t3":t3,
        "tema":tema,
        "trima":trima,
        "wma":wma,
        "adx":adx,
        "apo":apo,
        "aroondown":aroondown,
        "aroonup":aroonup,
        "bop":bop,
        "cci":cci,
        "cmo":cmo,
        "dx":dx,
        "mfi":mfi,
        "minus_di":minus_di,
        "mom":mom,
        "plus_di":plus_di,
        "ppo":ppo,
        "roc":roc,
        "rsi":rsi,
        "slowk":slowk,
        "slowd":slowd,
        "fastk":fastk,
        "fastd":fastd,
        "trix":trix,
        "ultosc":ultosc,
        "willr":willr,
        "ad":ad,
        "obv":obv,
        "atr":atr,
        "trange":trange,
        "beta":beta,
        "correl":correl,
        "linearreg":linearreg,
        "stddev":stddev,
        "tsf":tsf,
        "var":var
    })
    
    TI_DF.to_csv("../data/Stock_TI/"+symbol+"_TI.csv", index=False)
    print(symbol+" Technical Indicator Extraction!\nshape =",TI_DF.shape)

for symbol in NASDAQ_symbols: TI_ex(symbol, NASDAQ_stock)
